refactor the create person to module
