+++ student.rb
@@ -17,13 +17,13 @@ class Student < Person
     classroom.students.push(self) unless classroom.students.include?(self)
   end
 
-  def to_Json (*_args)
+  def to_json(*_args)
     {
-    JSON.create_id => self.class.name,
-    'id' => @id,
-    'age' => @age,
-    'name' => @name,
-    'parent_permission' => @parent_permission
+      JSON.create_id => self.class.name,
+      'id' => @id,
+      'age' => @age,
+      'name' => @name,
+      'parent_permission' => @parent_permission
     }
   end
 end
